requirements.txt
# rta-severity-predictor/requirements.txt
# IMPORTANT: versions must match the environment where models were trained

# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4

# Auth
python-jose[cryptography]==3.3.0
werkzeug==3.0.3

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
asyncpg==0.29.0

# ML Core — versions must match training environment exactly
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1

# Optional ML (install if available — app works without them)
# xgboost==2.0.3
# lightgbm==4.3.0
# imbalanced-learn==0.12.2
# shap==0.45.0
# optuna==3.6.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Testing
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
